# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CarbonPulse API — Precision carbon intelligence platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  region: string;
  weeklyBudgetKg: number;
  createdAt: string;
}

export interface UserInput {
  email: string;
  displayName?: string;
  avatarUrl?: string;
  region?: string;
}

export interface UserUpdate {
  displayName?: string;
  avatarUrl?: string;
  region?: string;
  weeklyBudgetKg?: number;
}

export interface Activity {
  id: string;
  userId: string;
  category: string;
  subcategory: string;
  quantity: number;
  unit: string;
  co2Grams: number;
  /** @nullable */
  description?: string | null;
  loggedAt: string;
  createdAt: string;
  /** @nullable */
  originLat?: number | null;
  /** @nullable */
  originLng?: number | null;
  /** @nullable */
  destLat?: number | null;
  /** @nullable */
  destLng?: number | null;
}

export interface ActivityInput {
  category: string;
  subcategory: string;
  quantity: number;
  unit: string;
  description?: string;
  loggedAt?: string;
  originLat?: number;
  originLng?: number;
  destLat?: number;
  destLng?: number;
}

export interface ActivityUpdate {
  quantity?: number;
  description?: string;
  loggedAt?: string;
}

export interface ActivityPage {
  data: Activity[];
  total: number;
  page: number;
  limit: number;
}

export interface CategoryTotal {
  category: string;
  co2Kg: number;
  count: number;
  percent: number;
}

export interface ActivitySummary {
  todayKg: number;
  weekKg: number;
  monthKg: number;
  weekBudgetKg: number;
  weekPercentUsed: number;
  /** @nullable */
  topCategory: string | null;
  categoryBreakdown: CategoryTotal[];
  recentActivities?: Activity[];
}

export interface TrendPoint {
  date: string;
  co2Kg: number;
}

export interface HeatmapDay {
  date: string;
  co2Kg: number;
  level: number;
}

export interface TransportPreviewInput {
  originLat: number;
  originLng: number;
  destLat: number;
  destLng: number;
}

export interface TransportModeEstimate {
  mode: string;
  label: string;
  co2Grams: number;
  emoji: string;
}

export interface TransportPreview {
  distanceKm: number;
  modes: TransportModeEstimate[];
}

export interface Insight {
  id: string;
  userId: string;
  weekStartDate: string;
  insightText: string;
  primaryDriver: string;
  suggestedAction: string;
  projectedSavingKgYear: number;
  totalCo2KgForWeek: number;
  deltaFromPrevWeek: number;
  createdAt: string;
}

export interface Goal {
  id: string;
  userId: string;
  targetKgPerWeek: number;
  /** @nullable */
  category?: string | null;
  isActive: boolean;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
}

export interface GoalInput {
  targetKgPerWeek: number;
  category?: string;
  startDate?: string;
  endDate?: string;
}

export interface GoalUpdate {
  targetKgPerWeek?: number;
  isActive?: boolean;
  endDate?: string;
}

export interface GoalProgress {
  hasGoal: boolean;
  targetKgPerWeek: number;
  currentKg: number;
  remainingKg: number;
  percentUsed: number;
  onTrack: boolean;
}

export interface CategoryAverage {
  category: string;
  avgKg: number;
  userKg: number;
}

export interface CommunityStats {
  userWeekKg: number;
  globalAvgKg: number;
  userPercentile: number;
  totalUsers: number;
  regionLabel: string;
  categoryAverages: CategoryAverage[];
}

export interface EmissionFactor {
  id: string;
  category: string;
  subcategory: string;
  factorGPerUnit: number;
  unit: string;
  label: string;
  source: string;
}

export interface Badge {
  id: string;
  label: string;
  description: string;
  earned: boolean;
  /** @nullable */
  earnedAt: string | null;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  todayLogged: boolean;
  weeklyBudgetKg: number;
  weekKg: number;
  badges: Badge[];
}

export type ListActivitiesParams = {
page?: number;
limit?: number;
category?: string;
from?: string;
to?: string;
};

export type ListInsightsParams = {
limit?: number;
};