# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CarbonPulse API — Precision carbon intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List user activities with optional filters
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: category
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Paginated activity list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityPage"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log a new carbon activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created activity with calculated CO2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
...
  /activities/transport-preview:
    post:
      operationId: previewTransport
      tags: [activities]
      summary: Preview CO2 for a transport route before logging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportPreviewInput"
      responses:
        "200":
          description: Route distance and CO2 estimates per mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportPreview"
  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get a single activity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── INSIGHTS ──────────────────────────────────────────────────────────────────
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List user goals
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a carbon reduction goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── COMMUNITY ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Activity:
      type: object
      required: [id, userId, category, subcategory, quantity, unit, co2Grams, loggedAt, createdAt]
      properties:
        id:          { type: string }
        userId:      { type: string }
        category:    { type: string }
        subcategory: { type: string }
        quantity:    { type: number }
        unit:        { type: string }
        co2Grams:    { type: number }
        description: { type: ["string", "null"] }
        loggedAt:    { type: string }
        createdAt:   { type: string }
        originLat:   { type: ["number", "null"] }
        originLng:   { type: ["number", "null"] }
        destLat:     { type: ["number", "null"] }
        destLng:     { type: ["number", "null"] }
    ActivityInput:
      type: object
      required: [category, subcategory, quantity, unit]
      properties:
        category:    { type: string }
        subcategory: { type: string }
        quantity:    { type: number }
        unit:        { type: string }
        description: { type: string }
        loggedAt:    { type: string }
        originLat:   { type: number }
        originLng:   { type: number }
        destLat:     { type: number }
        destLng:     { type: number }
    ActivityUpdate:
      type: object
      properties:
        quantity:    { type: number }
        description: { type: string }
        loggedAt:    { type: string }
    ActivityPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Activity" } }
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
...
    TransportPreviewInput:
      type: object
      required: [originLat, originLng, destLat, destLng]
      properties:
        originLat: { type: number }
        originLng: { type: number }
        destLat:   { type: number }
        destLng:   { type: number }
    TransportPreview:
      type: object
      required: [distanceKm, modes]
      properties:
        distanceKm: { type: number }
        modes:
          type: array
          items:
            $ref: "#/components/schemas/TransportModeEstimate"
    TransportModeEstimate:
      type: object
      required: [mode, label, co2Grams, emoji]
      properties:
        mode:     { type: string }
        label:    { type: string }
        co2Grams: { type: number }
        emoji:    { type: string }

    # ── INSIGHT ────────────────────────────────────────────────────────────────
...
    Goal:
      type: object
      required: [id, userId, targetKgPerWeek, isActive, startDate, createdAt]
      properties:
        id:               { type: string }
        userId:           { type: string }
        targetKgPerWeek:  { type: number }
        category:         { type: ["string", "null"] }
        isActive:         { type: boolean }
        startDate:        { type: string }
        endDate:          { type: ["string", "null"] }
        createdAt:        { type: string }
    GoalInput:
      type: object
      required: [targetKgPerWeek]
      properties:
        targetKgPerWeek: { type: number }
        category:        { type: string }
        startDate:       { type: string }
        endDate:         { type: string }
    GoalUpdate:
      type: object
      properties:
        targetKgPerWeek: { type: number }
        isActive:        { type: boolean }
        endDate:         { type: string }
...